edit scores3.c

+++ scores3.c
@@ -26,7 +26,7 @@ float average(int length, int array[])
         {
             sum += array[i];
         }
-        return sum / length;
+        return (float) sum / (float) length;
     }
